backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery>=5.3.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
decouple==0.0.7
defusedxml==0.8.0rc2
dj-database-url==2.0.0
Django==4.2
django-cors-headers==4.0.0
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.3.1
dotenv==0.0.5
drf-nested-routers==0.94.1
drf-yasg==1.21.8
fonttools==4.55.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kiwisolver==1.4.7
kombu==5.4.2
matplotlib==3.10.0
msgpack==1.1.0
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.52
Pillow==9.5.0
prometheus-client==0.19.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sentry-sdk==2.19.2
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.4.0
zope.interface==7.2
